TTS_API/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==1.10.15
google-cloud-texttospeech==2.16.4
slowapi==0.1.9
orjson==3.10.7

